Optimize node install by checking before command
Currently the repository setup steps rely shelling out to the nodejs
installer. Ansible has no way of automatically knowing if this needs to
be done so it was running it every time.

Manually added a check for the resulting tools we want to allow this
step to be skipped in future runs.